Move Study/Practice UI components to app/ui/learn and repoint imports
