+++ app/ui/learn/chat/chat-input.tsx
