+++ app/ui/learn/exam/exam-wrapper.tsx
@@ -6,11 +6,11 @@ import { useSearchParams, usePathname, useRouter } from 'next/navigation'
 import { ChatBubbleOvalLeftEllipsisIcon } from '@heroicons/react/24/outline'
 import useExplain from '@/app/hooks/useExplain'
 
-import ProgressWrapper from '@/app/ui/practice/exam/progress-wrapper'
-import Card from '@/app/ui/practice/exam/card'
-import NavButtons from '@/app/ui/practice/exam/nav-buttons'
-import Review from '@/app/ui/practice/exam/review'
-import Drawer from '@/app/ui/practice/exam/drawer'
+import ProgressWrapper from '@/app/ui/learn/exam/progress-wrapper'
+import Card from '@/app/ui/learn/exam/card'
+import NavButtons from '@/app/ui/learn/exam/nav-buttons'
+import Review from '@/app/ui/learn/exam/review'
+import Drawer from '@/app/ui/learn/exam/drawer'
 
 import { examMetadata } from '@/app/lib/seed'
 import { fetchEnhancedBundle } from '@/app/lib/explain'
